get finger gesture in a class
